+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 
-import Login from "@/app/pages/login";
+import Login from "@/app/login/page";
 
 export default function Home() {
